Support range of cron versions
