+++ ractor_actors/Cargo.toml
@@ -15,7 +15,12 @@ categories = ["asynchronous"]
 [features]
 filewatcher = ["notify"]
 net = ["tokio/net", "tokio/io-util", "tokio/macros", "tokio-rustls"]
-time = ["chrono", "cron", "dep:async-trait"]
+time = ["cron-0-12"] # deprecated
+cron-0-12 = ["dep:async-trait", "dep:chrono", "dep:cron-0-12"]
+cron-0-13 = ["dep:async-trait", "dep:chrono", "dep:cron-0-13"]
+cron-0-14 = ["dep:async-trait", "dep:chrono", "dep:cron-0-14"]
+cron-0-15 = ["dep:async-trait", "dep:chrono", "dep:cron-0-15"]
+cron-0-16 = ["dep:async-trait", "dep:chrono", "dep:cron-0-16"]
 streams = ["tokio-stream", "dep:async-trait"]
 sync = []
 watchdog = []
@@ -31,7 +36,11 @@ tracing = "0.1"
 # Feature-specific dependencies
 async-trait = { version = "0.1", optional = true }
 chrono = { version = "0.4", optional = true }
-cron = { version = "0.12", optional = true }
+cron-0-12 = { package = "cron", version = "0.12", optional = true }
+cron-0-13 = { package = "cron", version = "0.13", optional = true }
+cron-0-14 = { package = "cron", version = "0.14", optional = true }
+cron-0-15 = { package = "cron", version = "0.15", optional = true }
+cron-0-16 = { package = "cron", version = "0.16", optional = true }
 tokio-stream = { version = "0.1", optional = true }
 notify = { version = "8", optional = true }
 tokio-rustls = { version = "0.26", optional = true }
@@ -43,5 +52,12 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 tracing-test = "0.2"
 
 [[example]]
+name = "cron"
+required-features = ["cron-0-16"]
+
+[[example]]
 name = "tcp_client"
 required-features = ["tokio/io-std", "watchdog"]
+
+[package.metadata.docs.rs]
+features = ["cron-0-12", "cron-0-13", "cron-0-14", "cron-0-15", "cron-0-16"]
