+++ ractor_actors/src/time/cron/mod.rs
@@ -18,9 +18,9 @@
 //! use std::str::FromStr;
 //! use std::time::Duration;
 //!
-//! use cron::Schedule;
+//! use cron_0_16::Schedule;
 //! use ractor::{Actor, ActorProcessingErr};
-//! use ractor_actors::time::cron::*;
+//! use ractor_actors::time::cron_0_16::*;
 //!
 //! struct SomeJob;
 //!
@@ -67,38 +67,12 @@
 
 use std::collections::HashMap;
 
-use cron::Schedule;
-use ractor::{Actor, ActorId, ActorProcessingErr, ActorRef, RpcReplyPort, State, SupervisionEvent};
+use super::cron_crate::Schedule;
+use super::{CronEventSubscriber, Job};
+use ractor::{Actor, ActorId, ActorProcessingErr, ActorRef, RpcReplyPort, SupervisionEvent};
 
 mod worker;
 use worker::{Cron, CronMessage};
-
-/// Represents a job managed by a cron schedule. Executes on a
-/// given period and may take an unknown amount of time.
-///
-/// If the job takes longer than the period, queueing may occur and the
-/// job may violate scheduling
-#[async_trait::async_trait]
-pub trait Job: State {
-    /// Retrieve the name of the cron job for logging
-    fn id<'a>(&self) -> &'a str;
-
-    /// Execute the work, taking a mutable reference to the object
-    async fn work(&mut self) -> Result<(), ActorProcessingErr>;
-}
-
-/// Represents a dynamic subscription to [CronManager] events which
-/// denote cron job startup/failure/stoppage
-pub trait CronEventSubscriber: State {
-    /// A job started
-    fn started(&self, job: String);
-
-    /// A job was stopped, with the optional stop reason
-    fn stopped(&self, job: String, reason: Option<String>);
-
-    /// A job failed, including the failure reason
-    fn failed(&self, job: String, reason: String);
-}
 
 /// The settings for a singular cron job
 pub struct CronSettings {
